WebScraper3.0/requirements.txt
selenium
beautifulsoup4
googlesearch-python
langchain
langchain-ollama
openai
python-dotenv
firecrawl
cohere
groq
pandas
streamlit
rich
datetime
requests